backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
black==23.3.0
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-nextjs==2.2.2
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
djangorestframework-camel-case==1.4.2
executing==1.2.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
install==1.3.5
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
model-bakery==1.11.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2023.3
requests==2.28.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==11.0.1
yarl==1.8.2
